refactor: Organize code
