+++ .env.sample
@@ -77,7 +77,7 @@ PHRASE_TIMEOUT=2
 # Useful if you do not want old subtitles to suddenly appear and overwrite current ones
 REQUEST_TIMEOUT=4
 
-# Language to translate from
+# Language to translate from or for tts to speak in
 # Use this website to select the correct language code according to ISO 639-1 https://www.andiamo.co.uk/resources/iso-language-codes/
 TARGET_LANGUAGE_CODE=ja
