+++ src/modules/audio_translate.py
@@ -1,6 +1,7 @@
 from datetime import datetime, timedelta
 from io import BytesIO
 from os import getenv
+from pathlib import Path
 from queue import Queue
 from threading import Thread
 from time import sleep
@@ -15,7 +16,7 @@ RECORD_TIMEOUT = int(getenv('RECORD_TIMEOUT'))
 PHRASE_TIMEOUT = int(getenv('PHRASE_TIMEOUT'))
 INPUT_LANGUAGE = getenv('TARGET_LANGUAGE_CODE')
 LOGGING = getenv("LOGGING", 'False').lower() in ('true', '1', 't')
-APP_AUDIO_WAV_PATH = r'audio\app_audio.wav'
+APP_AUDIO_WAV_PATH = Path(__file__).resolve().parent.parent / r'audio\app_audio.wav'
 
 
 def request_thread(queue, phrase_time, now):
