+++ src/modules/voicevox.py
@@ -4,19 +4,15 @@ from threading import Thread
 from urllib.parse import urlencode
 
 import requests
-import sounddevice as sd
-import soundfile as sf
 from dotenv import load_dotenv
-import keyboard
+
+from modules.audio_to_device import play_voice
 
 load_dotenv()
 
 # Audio devices
 SPEAKERS_INPUT_ID = int(getenv('VOICEMEETER_INPUT_ID'))
 APP_INPUT_ID = int(getenv('CABLE_INPUT_ID'))
-
-# Keyboard
-INGAME_PUSH_TO_TALK_KEY = getenv('INGAME_PUSH_TO_TALK_KEY')
 
 # Voicevox settings
 BASE_URL = getenv('VOICEVOX_BASE_URL')
@@ -26,20 +22,8 @@ VOLUME_SCALE = float(getenv('VOLUME_SCALE'))
 INTONATION_SCALE = float(getenv('INTONATION_SCALE'))
 PRE_PHONEME_LENGTH = float(getenv('PRE_PHONEME_LENGTH'))
 POST_PHONEME_LENGTH = float(getenv('POST_PHONEME_LENGTH'))
-VOICEVOX_WAV_PATH = Path(__file__).resolve().parent.parent / r'audio\voicevox.wav'
 
-
-def play_voice(device_id):
-    data, fs = sf.read(VOICEVOX_WAV_PATH, dtype='float32')
-
-    if INGAME_PUSH_TO_TALK_KEY:
-        keyboard.press(INGAME_PUSH_TO_TALK_KEY)
-
-    sd.play(data, fs, device=device_id)
-    sd.wait()
-
-    if INGAME_PUSH_TO_TALK_KEY:
-        keyboard.release(INGAME_PUSH_TO_TALK_KEY)
+TTS_WAV_PATH = Path(__file__).resolve().parent.parent / r'audio\tts.wav'
 
 
 def speak_jp(sentence):
@@ -62,21 +46,7 @@ def speak_jp(sentence):
     params_encoded = urlencode({'speaker': VOICE_ID})
     r = requests.post(f'{BASE_URL}/synthesis?{params_encoded}', json=voicevox_query)
 
-    with open(VOICEVOX_WAV_PATH, 'wb') as outfile:
-        outfile.write(r.content)
-
-    # play voice to app mic input and speakers/headphones
-    threads = [Thread(target=play_voice, args=[APP_INPUT_ID]), Thread(target=play_voice, args=[SPEAKERS_INPUT_ID])]
-    [t.start() for t in threads]
-    [t.join() for t in threads]
-
-
-def speak_de(sentence):
-    # generate initial query
-    params = {'text': sentence}
-    r = requests.get('http://localhost:50021/api/tts', params=params)
-
-    with open(VOICEVOX_WAV_PATH, 'wb') as outfile:
+    with open(TTS_WAV_PATH, 'wb') as outfile:
         outfile.write(r.content)
 
     # play voice to app mic input and speakers/headphones
