+++ src/voice_translator.py
@@ -1,5 +1,6 @@
 import wave
 from os import getenv
+from pathlib import Path
 from time import sleep
 
 import deepl
@@ -20,7 +21,7 @@ TARGET_LANGUAGE = getenv('TARGET_LANGUAGE_CODE')
 MIC_ID = int(getenv('MICROPHONE_ID'))
 RECORD_KEY = getenv('MIC_RECORD_KEY')
 LOGGING = getenv('LOGGING', 'False').lower() in ('true', '1', 't')
-MIC_AUDIO_PATH = r'audio/mic.wav'
+MIC_AUDIO_PATH = Path(__file__).resolve().parent / r'audio/mic.wav'
 CHUNK = 1024
 FORMAT = pyaudio.paInt16
 
@@ -51,7 +52,7 @@ def on_release_key(_):
         return
 
     # write microphone audio to file
-    wf = wave.open(MIC_AUDIO_PATH, 'wb')
+    wf = wave.open(str(MIC_AUDIO_PATH), 'wb')
     wf.setnchannels(MIC_CHANNELS)
     wf.setsampwidth(p.get_sample_size(FORMAT))
     wf.setframerate(MIC_SAMPLING_RATE)
@@ -68,15 +69,15 @@ def on_release_key(_):
     if eng_speech:
 
         if USE_DEEPL:
-            final_speech = translator.translate_text(eng_speech, target_lang=TARGET_LANGUAGE)
+            translated_speech = translator.translate_text(eng_speech, target_lang=TARGET_LANGUAGE)
         else:
-            final_speech = translator.translate(eng_speech, dest=TARGET_LANGUAGE).text
+            translated_speech = translator.translate(eng_speech, dest=TARGET_LANGUAGE).text
 
         if LOGGING:
             print(f'English: {eng_speech}')
-            print(f'Translated: {final_speech}')
+            print(f'Translated: {translated_speech}')
 
-        speak(final_speech, TARGET_LANGUAGE)
+        speak(translated_speech, TARGET_LANGUAGE)
 
     else:
         print('No speech detected.')
